some more cleanup
