+++ pkg/network/tracer/connection/tracer_linux.go
@@ -1,0 +1,22 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build linux_bpf
+
+package connection
+
+import (
+	"github.com/DataDog/datadog-agent/comp/core/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/network/config"
+)
+
+// NewTracer returns a new Tracer
+func NewTracer(cfg *config.Config, telemetryComp telemetry.Component) (Tracer, error) {
+	if cfg.EnableEbpfless {
+		return newEbpfLessTracer(cfg)
+	}
+
+	return newEbpfTracer(cfg, telemetryComp)
+}
